+++ js/util.js
@@ -1,0 +1,4 @@
+export const getRandomInteger = (min, max) => {
+  const rand = min - 0.5 + Math.random() * (max - min + 1);
+  return Math.round(rand);
+};
